CLI_TOOL/requirements.txt
asgiref==3.2.5
certifi==2019.11.28
Distance==0.1.3
Django==3.0.4
image==1.5.28
inexactsearch==1.0.2
numpy==1.18.1
nltk==3.4.5
opencv-python==4.2.0.32
Pillow==7.0.0
pyenchant==3.0.1
pymongo==3.10.1
pyspellchecker==0.5.4
pytesseract==0.3.3
pytz==2019.3
silpa-common==0.3
soundex==1.1.3
sqlparse==0.3.1
tensorflow==2.0.0
wincertstore==0.2